BC-2552 Move mapFileRecordToFileRecordParams from helper to mapper
